+++ cmd/demo/demo.go
@@ -33,7 +33,7 @@ import (
 	"time"
 	"unicode/utf8"
 
-	ctree "github.com/crdteam/causal-tree/crdt/causaltree"
+	"github.com/crdteam/causal-tree/crdt/causaltree"
 )
 
 var (
@@ -63,7 +63,7 @@ type debugMessage struct {
 
 type treeinfo struct {
 	id    string
-	site  *ctree.CausalTree
+	site  *causaltree.CausalTree
 	mu    *sync.Mutex
 	order int
 }
@@ -89,7 +89,7 @@ type state struct {
 }
 
 func newState(debugMsgs chan<- debugMessage) *state {
-	site := ctree.New()
+	site := causaltree.New()
 	siteID := site.SiteID.String()
 	tree := treeinfo{
 		id:    siteID,
@@ -453,12 +453,12 @@ func unlockAll(trees ...treeinfo) {
 
 // -----
 
-func (s *state) debugTrees() []*ctree.CausalTree {
+func (s *state) debugTrees() []*causaltree.CausalTree {
 	if !s.isDebug() {
 		return nil
 	}
 	treeinfos := s.treeinfos()
-	trees := make([]*ctree.CausalTree, len(treeinfos))
+	trees := make([]*causaltree.CausalTree, len(treeinfos))
 	for i, info := range treeinfos {
 		trees[i] = info.site
 	}
