--- crdt/crdt_test/testdata/fuzz/FuzzList/ctrlaltdel
@@ -1,3 +1,0 @@
-go test fuzz v1
-[]byte("\\x00\\x00\\x23\\x00\\x00\\x2d\\x00\\x00\\x24\\x05\\x00\\x05\\x01\\x01\\x00\\x01\\x00\\x00\\x00\\x34\\x00\\x00\\x32\\x00\\x00\\x2c\\x00\\x01\\x21\\x00\\x01\\x2c\\x00\\x01\\x34\\x00\\x02\\x24\\x00\\x02\\x25\\x00\\x02\\x2c\\x06\\x00\\x01\\x06\\x02\\x01\\x06\\x02\\x00")
-
