--- crdt/crdt_test/testdata/fuzz/FuzzList/xabdy
@@ -1,3 +1,0 @@
-go test fuzz v1
-[]byte("\\x00\\x00\\x41\\x00\\x00\\x42\\x00\\x00\\x43\\x00\\x00\\x44\\x05\\x00\\x03\\x01\\x58\\x00\\x04\\x01\\x04\\x03\\x01\\x59\\x04\\x02\\x01\\x05\\x00\\x00\\x45\\x00\\x00\\x46\\x00\\x00\\x47\\x06\\x01\\x00\\x06\\x00\\x01\\x00\\x01\\x5a")
-
